feat: add user authentication and authorization, implement user routes and controllers, and configure environment variables

+++ controllers/actividad.js
@@ -3,6 +3,10 @@ import { ActividadModel } from "../models/actividad.js"
 
 export class ActividadController {
     static getAll = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+
         try{
         const actividades = await ActividadModel.getAll()
         res.json(actividades)
@@ -14,6 +18,10 @@ export class ActividadController {
     }
     
     static getById = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         const actividad = await ActividadModel.getById(id)
@@ -25,6 +33,10 @@ export class ActividadController {
         }
     }
     static postActividad = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+
         try {
             const validated_input = actividadSchema.parse(req.body);
             await ActividadModel.postActividad(validated_input)
@@ -35,6 +47,10 @@ export class ActividadController {
         }
     }
     static deleteActividad = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         await ActividadModel.deleteById(id)
@@ -45,6 +61,10 @@ export class ActividadController {
         }
     }
     static updateActividad = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+            
         try{
         const { id } = req.params
         const validated_input = actividadSchema.parse(req.body);
