+++ controllers/espacio.js
@@ -1,8 +1,11 @@
-import {EspacioSchema} from "../schemas/espacio.js"
-import { CategoriaSchema } from "../schemas/espacio.js"
+import {CategoriasSchema, EspacioSchema, CategoriaSchema} from "../schemas/espacio.js"
 import {EspacioModel, CategoriaModel} from "../models/espacio.js"
 export class EspacioController {
     static getAll = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const espacios = await EspacioModel.getAll()
         res.json(espacios)
@@ -14,6 +17,10 @@ export class EspacioController {
     }
     
     static getById = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         const espacios = await EspacioModel.getById(id)
@@ -25,8 +32,15 @@ export class EspacioController {
         }
     }
     static postEspacio = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try {
             const validated_input = EspacioSchema.parse(req.body);
+             if (req.file) {
+                validated_input.imagen = `/uploads/${req.file.filename}`;
+             }
             await EspacioModel.postEspacio(validated_input)
             res.status(201).json({"ok": true}).end(); 
         } catch (error) {
@@ -35,6 +49,10 @@ export class EspacioController {
         }
     }
     static deleteEspacio = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         await EspacioModel.deleteById(id)
@@ -45,11 +63,41 @@ export class EspacioController {
         }
     }
     static updateEspacio = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         const validated_input = EspacioSchema.parse(req.body);
+        if (req.file) {
+            //Obtener el espacio actual de la BD
+            const espacioActual = await EspacioModel.getById(id);
+            //Borrar la imagen anterior del servidor si existe
+            if (espacioActual && espacioActual.imagen) {
+                const oldPath = path.join(process.cwd(), espacioActual.imagen);
+            //Guardar la nueva ruta
+            if (fs.existsSync(oldPath)) fs.unlinkSync(oldPath);
+            }
+            validated_input.imagen = `/uploads/${req.file.filename}`;
+        }
         await EspacioModel.updateEspacio(id, validated_input)
         res.status(200).json({"ok": true}).end()
+        } catch (error) {
+            console.error(error);
+            res.status(400).json({ error: error.message });
+        }
+    }
+    static addCategoria = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
+        try {
+            const { id } = req.params
+            const validated_input = CategoriasSchema.parse(req.body); 
+            await EspacioModel.addCategorias(id, validated_input)
+            res.status(201).json({"ok": true}).end(); 
         } catch (error) {
             console.error(error);
             res.status(400).json({ error: error.message });
@@ -58,6 +106,10 @@ export class EspacioController {
 }
 export class CategoriaController {
     static getAll = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const categorias = await CategoriaModel.getAll()
         res.json(categorias)
@@ -69,6 +121,10 @@ export class CategoriaController {
     }
     
     static getById = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         const categoria = await CategoriaModel.getById(id)
@@ -80,6 +136,10 @@ export class CategoriaController {
         }
     }
     static postCategoria = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try {
             const validated_input = CategoriaSchema.parse(req.body);
             await CategoriaModel.postCategoria(validated_input)
@@ -90,6 +150,10 @@ export class CategoriaController {
         }
     }
     static deleteCategoria = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+
         try{
         const { id } = req.params
         await CategoriaModel.deleteById(id)
@@ -100,6 +164,10 @@ export class CategoriaController {
         }
     }
     static updateCategoria = async (req, res) => {
+        //Validamos la session
+        const { user } = req.session
+        if(!user || !user.administrador) return res.status(403).send('Access not authorized')
+            
         try{
         const { id } = req.params
         const validated_input = CategoriaSchema.parse(req.body);
